+++ apps/frontend/src/api/general/useServerGetInfo.ts
@@ -1,11 +1,11 @@
+import { api } from "@/config/client";
 import { useQuery } from "@tanstack/react-query";
-import { useAPI } from "./useAPI";
 
 export function useServerGetInfo() {
-  const { getServerInfo } = useAPI();
+  
   return useQuery({
     queryKey: ["server"],
-    queryFn: () => getServerInfo(),
+    queryFn: ()=>api.get("/").then(res=>res.data as {version: string}),
     staleTime: 1000 * 60 * 5,
   });
 }