+++ apps/frontend/src/components/custom/ServerReady.tsx
@@ -1,11 +1,11 @@
-import { useAuth } from "@/hooks/useAuth";
+import { useAuth } from "@/context/AuthProvider";
 import { Loader } from "./Loader";
 
 export function ServerReady({
   children,
 }: Readonly<{ children: React.ReactNode }>) {
   const { isServerReady } = useAuth();
-  console.log(isServerReady);
+
   if (!isServerReady) {
     return (
       <div className="flex items-center justify-center min-h-screen bg-gray-100">