+++ apps/frontend/src/main.tsx
@@ -3,6 +3,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { RouterProvider } from "react-router";
 import { ServerReady } from "./components/custom/ServerReady";
+import { Toaster } from "./components/ui/toaster";
 import { AuthProvider } from "./context/AuthProvider";
 import "./index.css";
 import { router } from "./router";
@@ -17,6 +18,7 @@ createRoot(document.getElementById("root")!).render(
           <RouterProvider router={router} />
         </ServerReady>
       </AuthProvider>
+      <Toaster />
     </QueryClientProvider>
   </StrictMode>,
 );