+++ apps/frontend/src/pages/Login.tsx
@@ -9,8 +9,9 @@ import {
 } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { AUTH } from "@/services/auth";
-import { useState, type FormEvent } from "react";
+import { useAuth } from "@/context/AuthProvider";
+import { AuthenticationService } from "@/services/AuthenticationService";
+import { useEffect, useState, type FormEvent } from "react";
 import { useNavigate } from "react-router";
 
 export default function LoginPage() {
@@ -18,7 +19,12 @@ export default function LoginPage() {
   const [password, setPassword] = useState("");
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState("");
+  const { isLogged } = useAuth();
   const navigate = useNavigate();
+
+  useEffect(() => {
+    if (isLogged) navigate("/me");
+  }, [isLogged, navigate]);
 
   const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
@@ -38,7 +44,7 @@ export default function LoginPage() {
       return;
     }
 
-    AUTH.login({ email, password })
+    AuthenticationService.login({ email, password })
       .then((res) => {
         if (res.error) {
           setError(res.error.message);